// autogpt_platform/frontend/src/app/api/__generated__/models/storeListingWithVersionsAdminView.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { StoreListingWithVersionsAdminViewActiveListingVersionId } from "./storeListingWithVersionsAdminViewActiveListingVersionId";
import type { StoreListingWithVersionsAdminViewCreatorEmail } from "./storeListingWithVersionsAdminViewCreatorEmail";
import type { StoreListingWithVersionsAdminViewLatestVersion } from "./storeListingWithVersionsAdminViewLatestVersion";
import type { StoreSubmissionAdminView } from "./storeSubmissionAdminView";

/**
 * A store listing with its version history
 */
export interface StoreListingWithVersionsAdminView {
  listing_id: string;
  graph_id: string;
  slug: string;
  active_listing_version_id?: StoreListingWithVersionsAdminViewActiveListingVersionId;
  has_approved_version?: boolean;
  creator_email?: StoreListingWithVersionsAdminViewCreatorEmail;
  latest_version?: StoreListingWithVersionsAdminViewLatestVersion;
  versions?: StoreSubmissionAdminView[];
}