// autogpt_platform/frontend/src/app/api/__generated__/models/storeSubmissionAdminView.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { StoreSubmissionAdminViewInstructions } from "./storeSubmissionAdminViewInstructions";
import type { StoreSubmissionAdminViewVideoUrl } from "./storeSubmissionAdminViewVideoUrl";
import type { StoreSubmissionAdminViewAgentOutputDemoUrl } from "./storeSubmissionAdminViewAgentOutputDemoUrl";
import type { StoreSubmissionAdminViewSubmittedAt } from "./storeSubmissionAdminViewSubmittedAt";
import type { StoreSubmissionAdminViewChangesSummary } from "./storeSubmissionAdminViewChangesSummary";
import type { SubmissionStatus } from "./submissionStatus";
import type { StoreSubmissionAdminViewReviewedAt } from "./storeSubmissionAdminViewReviewedAt";
import type { StoreSubmissionAdminViewReviewerId } from "./storeSubmissionAdminViewReviewerId";
import type { StoreSubmissionAdminViewReviewComments } from "./storeSubmissionAdminViewReviewComments";
import type { StoreSubmissionAdminViewInternalComments } from "./storeSubmissionAdminViewInternalComments";

export interface StoreSubmissionAdminView {
  listing_id: string;
  user_id: string;
  slug: string;
  listing_version_id: string;
  listing_version: number;
  graph_id: string;
  graph_version: number;
  name: string;
  sub_heading: string;
  description: string;
  instructions: StoreSubmissionAdminViewInstructions;
  categories: string[];
  image_urls: string[];
  video_url: StoreSubmissionAdminViewVideoUrl;
  agent_output_demo_url: StoreSubmissionAdminViewAgentOutputDemoUrl;
  submitted_at: StoreSubmissionAdminViewSubmittedAt;
  changes_summary: StoreSubmissionAdminViewChangesSummary;
  status: SubmissionStatus;
  reviewed_at?: StoreSubmissionAdminViewReviewedAt;
  reviewer_id?: StoreSubmissionAdminViewReviewerId;
  review_comments?: StoreSubmissionAdminViewReviewComments;
  run_count?: number;
  review_count?: number;
  review_avg_rating?: number;
  internal_comments: StoreSubmissionAdminViewInternalComments;
}